moved find scope tests in with other tests
